move random files to separate files
