+++ src/jdom/bridge.ts
@@ -2,7 +2,7 @@ import { JDBus } from "./bus"
 import { JDClient } from "./client"
 import { CHANGE, PACKET_PROCESS, PACKET_SEND } from "./constants"
 import Packet from "./packet"
-import { randomDeviceId } from "./utils"
+import { randomDeviceId } from "./random"
 
 /**
  * A client that bridges received and sent packets to a parent iframe.
